requirements.txt
streamlit
langgraph
langchain-openai
tavily
supabase
python-dotenv
psycopg[binary]
